services/rateapp/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-CLI==0.4.0
psycopg2-binary==2.7.6.1
Flask-Testing==0.6.2
coverage==4.5.3
flake8===3.7.7
flask-debugtoolbar==0.10.1
flask-cors==3.0.7